Move download verification functions into buddy_check.py
